Add files via upload
This was the practice exam I was doing to show you I did study and did the 2021 practice exam on almost every data set you provided for us. I'm just very upset it didn't knit so hopefully the outputs or explanation makes sense on what i was trying to do.